# specs/002_api_specification.yaml
...
openapi: "3.1.0"
info:
  title: Cyber Chakra Forensics Admin Portal API
  version: "1.0.0"
  description: |
    REST API for the CMF Admin Portal and the public-facing desktop app endpoints.
    All admin endpoints require Bearer token authentication.
    All public endpoints require a valid license key (X-License-Key header).
servers:
  - url: https://cyberchakra.online
    description: Production
  - url: https://staging.cyberchakra.online
    description: Staging

# =============================================================================
# Security Schemes
# =============================================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin portal JWT token obtained from /api/v1/auth/login.

    LicenseKeyAuth:
      type: apiKey
      in: header
      name: X-License-Key
      description: License key for public desktop app endpoints.

  # ===========================================================================
  # Shared Schemas
  # ===========================================================================
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 25
        total:
          type: integer
          example: 142
        total_pages:
          type: integer
          example: 6
        has_next:
          type: boolean
        has_prev:
          type: boolean
...
    AdminUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        username:
          type: string
        display_name:
          type: string
        role:
          type: string
          enum: [super_admin, admin, support, viewer, analyst]
        is_active:
          type: boolean
        mfa_enabled:
          type: boolean
        last_login_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time

    # ---- Analytics ----
...
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    QueryPage:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    QueryPerPage:
      name: per_page
      in: query
      schema:
        type: integer
        default: 25
        minimum: 1
        maximum: 100
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/admin/users:
    get:
      tags: [Admin]
      summary: List admin users
      parameters:
        - $ref: '#/components/parameters/QueryPage'
        - $ref: '#/components/parameters/QueryPerPage'
        - name: role
          in: query
          schema:
            type: string
            enum: [super_admin, admin, support, viewer, analyst]
        - name: is_active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Admin users list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminUser'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
    post:
      tags: [Admin]
      summary: Create admin user
      description: "Requires super_admin role."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, username, password, display_name, role]
              properties:
                email:
                  type: string
                  format: email
                username:
                  type: string
                  minLength: 3
                  maxLength: 32
                password:
                  type: string
                  minLength: 12
                display_name:
                  type: string
                role:
                  type: string
                  enum: [admin, support, viewer, analyst]
      responses:
        "201":
          description: Admin user created
        "409":
          description: Email or username already taken
  /api/v1/admin/users/{id}:
    get:
      tags: [Admin]
      summary: Get admin user details
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        "200":
          description: Admin user details
    patch:
      tags: [Admin]
      summary: Update admin user
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name:
                  type: string
                role:
                  type: string
                  enum: [admin, support, viewer, analyst]
                is_active:
                  type: boolean
      responses:
        "200":
          description: Admin user updated
    delete:
      tags: [Admin]
      summary: Soft-delete admin user
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        "204":
          description: Admin user deleted
  /api/v1/admin/users/{id}/reset-password:
    post:
      tags: [Admin]
      summary: Force reset admin password
      description: "Requires super_admin role."
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [new_password]
              properties:
                new_password:
                  type: string
                  minLength: 12
      responses:
        "200":
          description: Password reset
...